+++ tests/AppLibrary.Tests/PaginationTests/PaginatedResultTests.cs
@@ -13,20 +13,18 @@ public class PaginatedResultTests
         var result = new PaginatedResult<string>(_items, itemCount,
             new PaginatedRequest(1, _items.Length));
 
-        Assert.Multiple(() =>
-        {
-            result.CurrentCount.Should().Be(_items.Length);
-            result.PageNumber.Should().Be(1);
-            result.PageSize.Should().Be(_items.Length);
-            result.TotalCount.Should().Be(itemCount);
-            result.TotalPages.Should().Be(1);
-            result.FirstItemIndex.Should().Be(1);
-            result.HasNextPage.Should().BeFalse();
-            result.HasPreviousPage.Should().BeFalse();
-            result.LastItemIndex.Should().Be(2);
-            result.PreviousPageNumber.Should().Be(1);
-            result.NextPageNumber.Should().Be(1);
-        });
+        using var scope = new AssertionScope();
+        result.CurrentCount.Should().Be(_items.Length);
+        result.PageNumber.Should().Be(1);
+        result.PageSize.Should().Be(_items.Length);
+        result.TotalCount.Should().Be(itemCount);
+        result.TotalPages.Should().Be(1);
+        result.FirstItemIndex.Should().Be(1);
+        result.HasNextPage.Should().BeFalse();
+        result.HasPreviousPage.Should().BeFalse();
+        result.LastItemIndex.Should().Be(2);
+        result.PreviousPageNumber.Should().Be(1);
+        result.NextPageNumber.Should().Be(1);
     }
 
     [Test]
@@ -36,20 +34,18 @@ public class PaginatedResultTests
         var result = new PaginatedResult<string>(_items, itemCount,
             new PaginatedRequest(2, _items.Length));
 
-        Assert.Multiple(() =>
-        {
-            result.CurrentCount.Should().Be(_items.Length);
-            result.PageNumber.Should().Be(2);
-            result.PageSize.Should().Be(_items.Length);
-            result.TotalCount.Should().Be(itemCount);
-            result.TotalPages.Should().Be(5);
-            result.FirstItemIndex.Should().Be(3);
-            result.HasNextPage.Should().BeTrue();
-            result.HasPreviousPage.Should().BeTrue();
-            result.LastItemIndex.Should().Be(4);
-            result.PreviousPageNumber.Should().Be(1);
-            result.NextPageNumber.Should().Be(3);
-        });
+        using var scope = new AssertionScope();
+        result.CurrentCount.Should().Be(_items.Length);
+        result.PageNumber.Should().Be(2);
+        result.PageSize.Should().Be(_items.Length);
+        result.TotalCount.Should().Be(itemCount);
+        result.TotalPages.Should().Be(5);
+        result.FirstItemIndex.Should().Be(3);
+        result.HasNextPage.Should().BeTrue();
+        result.HasPreviousPage.Should().BeTrue();
+        result.LastItemIndex.Should().Be(4);
+        result.PreviousPageNumber.Should().Be(1);
+        result.NextPageNumber.Should().Be(3);
     }
 
     [Test]